requirements.txt
openai>=1.1.0
python-dotenv>=1.0.0
xrpl-py>=2.0.0
aiohttp>=3.8.5
